// app/type.ts
export interface WeatherHourDataType {
  time: string; // "2025-08-07 13:00"
  time_epoch: number; // 1754551800
  temp_c: number;
  temp_f: number;
  is_day: 0 | 1;
  condition: {
    text: string; // "Mist"
    icon: string; // "//cdn.weatherapi.com/weather/64x64/day/143.png"
    code: number; // 1030
  };
  wind_mph: number;
  wind_kph: number;
  wind_degree: number;
  wind_dir: string;
  pressure_mb: number;
  pressure_in: number;
  precip_mm: number;
  precip_in: number;
  humidity: number;
  cloud: number;
  feelslike_c: number;
  feelslike_f: number;
  windchill_c: number;
  windchill_f: number;
  heatindex_c: number;
  heatindex_f: number;
  dewpoint_c: number;
  dewpoint_f: number;
  will_it_rain: number;
  chance_of_rain: number;
  will_it_snow: number;
  chance_of_snow: number;
  vis_km: number;
  vis_miles: number;
  gust_mph: number;
  gust_kph: number;
  uv: number;
  snow_cm: number;
}